+++ pkg/ruler/ruler.go
@@ -85,6 +85,13 @@ type Config struct {
 	// HTTP timeout duration when sending notifications to the Alertmanager.
 	NotificationTimeout time.Duration `yaml:"notification_timeout"`
 
+	// Max time to tolerate outage for restoring "for" state of alert.
+	OutageTolerance time.Duration `yaml:"for_outage_tolerance"`
+	// Minimum duration between alert and restored "for" state. This is maintained only for alerts with configured "for" time greater than grace period.
+	ForGracePeriod time.Duration `yaml:"for_grace_period"`
+	// Minimum amount of time to wait before resending an alert to Alertmanager.
+	ResendDelay time.Duration `yaml:"resend_delay"`
+
 	// Enable sharding rule groups.
 	EnableSharding   bool          `yaml:"enable_sharding"`
 	SearchPendingFor time.Duration `yaml:"search_pending_for"`
@@ -129,6 +136,9 @@ func (cfg *Config) RegisterFlags(f *flag.FlagSet) {
 	f.DurationVar(&cfg.FlushCheckPeriod, "ruler.flush-period", 1*time.Minute, "Period with which to attempt to flush rule groups.")
 	f.StringVar(&cfg.RulePath, "ruler.rule-path", "/rules", "file path to store temporary rule files for the prometheus rule managers")
 	f.BoolVar(&cfg.EnableAPI, "experimental.ruler.enable-api", false, "Enable the ruler api")
+	f.DurationVar(&cfg.OutageTolerance, "ruler.for-outage-tolerance", time.Hour, `Max time to tolerate outage for restoring "for" state of alert.`)
+	f.DurationVar(&cfg.ForGracePeriod, "ruler.for-grace-period", 10*time.Minute, `Minimum duration between alert and restored "for" state. This is maintained only for alerts with configured "for" time greater than grace period.`)
+	f.DurationVar(&cfg.ResendDelay, "ruler.resend-delay", time.Minute, `Minimum amount of time to wait before resending an alert to Alertmanager.`)
 }
 
 // Ruler evaluates rules.
@@ -515,13 +525,16 @@ func (r *Ruler) newManager(ctx context.Context, userID string) (*rules.Manager, 
 	reg = prometheus.WrapRegistererWithPrefix("cortex_", reg)
 	logger := log.With(r.logger, "user", userID)
 	opts := &rules.ManagerOptions{
-		QueryFunc:   r.queryFunc(r.cfg.EvaluationDelay),
-		ParseExpr:   r.parseExpr,
-		Context:     user.InjectOrgID(ctx, userID),
-		ExternalURL: r.alertURL,
-		NotifyFunc:  sendAlerts(notifier, r.alertURL.String()),
-		Logger:      logger,
-		Registerer:  reg,
+		QueryFunc:       r.queryFunc(r.cfg.EvaluationDelay),
+		ParseExpr:       r.parseExpr,
+		Context:         user.InjectOrgID(ctx, userID),
+		ExternalURL:     r.alertURL,
+		NotifyFunc:      sendAlerts(notifier, r.alertURL.String()),
+		Logger:          logger,
+		Registerer:      reg,
+		OutageTolerance: r.cfg.OutageTolerance,
+		ForGracePeriod:  r.cfg.ForGracePeriod,
+		ResendDelay:     r.cfg.ResendDelay,
 	}
 	app, hist := r.appendableHistory(userID, opts)
 	opts.Appendable = app